+++ Modules/KiwiEngine/KiwiPatcher.cpp
@@ -87,7 +87,7 @@ namespace kiwi
             return links;
         }
         
-        void Patcher::sendToObject(Object& object, const uint32_t inlet, std::vector<Atom> args)
+        void Patcher::sendToObject(Object& object, const size_t inlet, std::vector<Atom> args)
         {
             object.receive(inlet, args);
         }
